+++ rickled/cli/conv.py
@@ -4,22 +4,21 @@ from rickled.tools import Converter
 
 def conv(args):
     try:
+        output_type = args.OUTPUT_TYPE.lower() if args.OUTPUT_TYPE else 'yaml'
         if args.INPUT or args.INPUT_DIRECTORY:
             Converter(input_files=args.INPUT,
                       output_files=args.OUTPUT,
                       input_directory=args.INPUT_DIRECTORY,
-                      default_output_type=args.OUTPUT_TYPE,
+                      default_output_type=output_type,
                       verbose=args.VERBOSE).do_convert()
-        elif args.OUTPUT_TYPE:
+        else:
             data = sys.stdin.read()
 
             converted = Converter.convert_string(input_string=data,
                                                  input_type=args.INPUT_TYPE,
-                                                 output_type=args.OUTPUT_TYPE)
+                                                 output_type=output_type)
 
             print(converted)
-        else:
-            raise CLIError(message='Incorrect usage of CLI tool, refer to documentation', cli_tool=CLIError.CLITool.CONV)
 
     except Exception as exc:
         raise CLIError(message=str(exc), cli_tool=CLIError.CLITool.CONV)